+++ src/popup/pages/OptionsPage/index.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import OptionsPageLayout from '../../../layouts/OptionsPageLayout';
+
+function OptionsPage() {
+  return (
+    <OptionsPageLayout>
+      <OptionsForm />
+    </OptionsPageLayout>
+  );
+}
+
+export default OptionsPage;