+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Create OptionsPageLayout with header and body padding
 
 ## [0.1.0] - 2022-01-01
 ### Added